requirements.txt
openai>=1.0.0
praw
google-api-python-client
google-auth
google-cloud-firestore
python-dotenv
